webserver_tester/requirements.txt
requests
Pillow
tqdm 
colorama
